Add files via upload
The goal of this project is to create a music player algorithm, Play music player, search, repeat, add track, remove track, and stop the music are the components of this musical player application. We used the java programming language, the NetBeans platform, MYSQL for our database, and Visio for our flowcharts. The Project has 7 members.